% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.package.R
\name{unwrap.package}
\alias{unwrap.package}
\title{Unwrap an installed wrapper package}
\usage{
unwrap.package(package)
}
\arguments{
\item{package}{Name of the installed wrapper package to be unwrapped, given as a character string.}
}
\description{
Install the inner package inside an installed wrapper package. Note that 
this will overwrite the wrapper package if the package name used in the
DESCRIPTION file of the wrapper package is the same as the package name
used in the DESCRIPTION file of the inner package. In fact this is the  
recommended way of using a wrapper package.
}

