% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.package.R
\name{wrap.package}
\alias{wrap.package}
\title{Convert a regular R package folder into a wrapper package.}
\usage{
wrap.package(package)
}
\arguments{
\item{package}{Path to the package folder to be converted into a wrapper package.}
}
\description{
Convert a regular R package folder into a wrapper package.
}

