% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.package.R
\name{load.package}
\alias{load.package}
\title{Load an R package}
\usage{
load.package(package = ".")
}
\arguments{
\item{package}{A character string which can be either the name of an installed package or the
path to a source package. The default value is ".", which assumes the current
working directory to be the root directory of a source package. If an installed 
package with the given name exists, that package will be either unwrapped or
simply reloaded, depending on whether an inner package is found. Otherwise the
source package at the given path is first installed, then the newly installed
package is either unwrapped or reloaded.}
}
\description{
Load an R package
}

