% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrap.package.R
\name{unwrap.package}
\alias{unwrap.package}
\title{Unwrap an installed wrapper package}
\usage{
unwrap.package(package)
}
\arguments{
\item{package}{A character string which gives the name of an installed wrapper package.
The inner package inside this wrapper package will be installed. Note that 
the wrapper package will be over-written if it has the same name as its
be the case if the wrapper package was created by the wrap.package function.
This is the recommended way of using a wrapper package, which is intended to
be only a container for its inner package, it is no longer needed once it is
unwrapped and the inner package is installed.}
}
\description{
Unwrap an installed wrapper package
}

