% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.package.R
\name{wrap.package}
\alias{wrap.package}
\title{Convert a regular source package into a wrapper package}
\usage{
wrap.package(package = ".")
}
\arguments{
\item{package}{Path to the source package to be converted into a wrapper package.}
}
\description{
Convert a regular source package into a wrapper package
}

